+++ cpsc471-libmansys-frontend/src/views/CdView.tsx
@@ -69,7 +69,7 @@ export const CdView = (cd: string) => {
             console.log("add item");
             addItemInfo(item);
         }
-        else {
+        else if(selectedId === "Remove Item") {
             console.log("delete item");
             deleteItemInfo(item);
         }
